v1.1.4
- Added extra safety checks for the string helper functions in "Printer.hpp/cpp".

+++ Framework/Printer.hpp
@@ -7,8 +7,7 @@
 #include <filesystem>
 #include <cstdarg>
 
-// Printer Width Types
-enum class EWidthTypes : size_t
+enum class EWidthTypes : uint32_t
 {
 	None = 0,
 	Byte = 2,
@@ -22,18 +21,18 @@ enum class EWidthTypes : size_t
 namespace Printer
 {
 	void Empty(std::ostringstream& stream);
-	void FillRight(std::ostringstream& stream, char fill, size_t width);
-	void FillLeft(std::ostringstream& stream, char fill, size_t width);
-	void FillRight(std::ofstream& stream, char fill, size_t width);
-	void FillLeft(std::ofstream& stream, char fill, size_t width);
+	void FillRight(std::ostringstream& stream, char fill, uint32_t width);
+	void FillLeft(std::ostringstream& stream, char fill, uint32_t width);
+	void FillRight(std::ofstream& stream, char fill, uint32_t width);
+	void FillLeft(std::ofstream& stream, char fill, uint32_t width);
 	void ReplaceChars(std::string& baseStr, char oldChar, char newChar);
 	void ReplaceString(std::string& baseStr, const std::string& strToReplace, const std::string& replaceWithStr);
-	std::string Hex(uint64_t decimal, size_t width);
-	std::string Hex(uint64_t decimal, EWidthTypes width = EWidthTypes::None);
+	std::string Hex(uintptr_t decimal, uint32_t width);
+	std::string Hex(uintptr_t decimal, EWidthTypes width = EWidthTypes::None);
 	std::string Hex(void* pointer);
-	std::string Decimal(uint64_t hex, size_t width);
-	std::string Decimal(uint64_t hex, EWidthTypes width = EWidthTypes::None);
-	std::string Precision(float value, size_t precision);
+	std::string Decimal(uintptr_t hex, uint32_t width);
+	std::string Decimal(uintptr_t hex, EWidthTypes width = EWidthTypes::None);
+	std::string Precision(float value, uint32_t precision);
 	std::string ToUpper(std::string str);
 	std::string ToLower(std::string str);
